[UPDATE] 회원가입기능 오류 해결, AuthForm 과 Auth 를 합침
